Fix up dates (mostly) plus script example
For run,dbl-run,ichimoku.py fixed the date issue for assets that are 24/7. The dates line up right. Stocks,metals & etc that trade 5 days a week, there is still an offset problem with index so beware the dates on the chart. No easy way to get the index to line up for these, just an FYI.

scan.sh - An example script that generates one chart after another as a summary.